# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin API 文档",
        "title": "Gin Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "25076778@qq.com"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/api/v1/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录相关"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/refresh-token": {
            "post": {
                "description": "刷新token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录相关"
                ],
                "summary": "刷新token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "刷新Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.Token"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.ArgsErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Invalid arguments"
                }
            }
        },
        "errcode.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "errcode.SystemErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "fullName": {
                    "description": "姓名",
                    "type": "string"
                },
                "gender": {
                    "description": "性别 1=男 2=女",
                    "type": "integer"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1=启用 2=停用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userRoles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserRoles"
                    }
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.UserRoles": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色id",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userId": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
...
        "request.UserLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "v1.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/v1.Token"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "v1.Token": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "refreshTokenExpire": {
                    "type": "integer",
                    "example": 172800
                },
                "tokenExpire": {
                    "type": "integer",
                    "example": 7200
                }
            }
        }
    }
}